+++ apiary-ranger-metastore-plugin/src/main/java/com/expediagroup/apiary/extensions/rangerauth/listener/ApiaryRangerAuthPreEventListener.java
@@ -15,9 +15,7 @@
  */
 package com.expediagroup.apiary.extensions.rangerauth.listener;
 
-import java.util.Date;
-import java.util.Set;
-
+import com.google.common.collect.Sets;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.metastore.MetaStorePreEventListener;
 import org.apache.hadoop.hive.metastore.api.Database;
@@ -50,8 +48,8 @@ import org.apache.ranger.plugin.policyengine.RangerAccessResult;
 import org.apache.ranger.plugin.service.RangerBasePlugin;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import com.google.common.collect.Sets;
+import java.util.Date;
+import java.util.Set;
 
 enum HiveAccessType {
   NONE,
@@ -83,13 +81,13 @@ public class ApiaryRangerAuthPreEventListener extends MetaStorePreEventListener 
     plugin = new RangerBasePlugin("hive", "metastore");
     plugin.init();
     plugin.setResultProcessor(new RangerDefaultAuditHandler());
-    log.info("ApiaryRangerAuthPreEventListener created");
+    log.debug("ApiaryRangerAuthPreEventListener created");
   }
 
   public ApiaryRangerAuthPreEventListener(Configuration config, RangerBasePlugin plugin) throws HiveException {
     super(config);
     this.plugin = plugin;
-    log.info("ApiaryRangerAuthPreEventListener created");
+    log.debug("ApiaryRangerAuthPreEventListener created");
   }
 
   @Override