+++ apiary-ranger-metastore-plugin/src/test/java/com/expediagroup/apiary/extensions/rangerauth/policyproviders/ApiaryRangerAuthAllAccessPolicyProviderTest.java
@@ -15,6 +15,13 @@
  */
 package com.expediagroup.apiary.extensions.rangerauth.policyproviders;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.equalTo;
+import static org.hamcrest.Matchers.hasItems;
+import static org.hamcrest.Matchers.hasSize;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+
 import com.expediagroup.apiary.extensions.rangerauth.listener.RangerAdminClientImpl;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.ql.security.authorization.plugin.HiveOperationType;
@@ -27,7 +34,6 @@ import org.apache.ranger.plugin.policyengine.RangerAccessResourceImpl;
 import org.apache.ranger.plugin.policyengine.RangerAccessResult;
 import org.apache.ranger.plugin.policyengine.RangerPolicyEngineImpl;
 import org.apache.ranger.plugin.util.ServicePolicies;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
@@ -37,17 +43,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 import java.util.Date;
 import java.util.Set;
 
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.hasItems;
-import static org.hamcrest.Matchers.hasSize;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.anyLong;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.MockitoAnnotations.initMocks;
-
 @RunWith(MockitoJUnitRunner.class)
 public class ApiaryRangerAuthAllAccessPolicyProviderTest {
   @Mock
@@ -55,14 +50,9 @@ public class ApiaryRangerAuthAllAccessPolicyProviderTest {
   @Mock
   private Configuration configuration;
 
-  private static final String serviceName = "apiary";
-  private static final String appId = "appid";
-  private static final String configPropertyPrefix = "prefix";
-
-  @Before
-  public void setup() {
-    initMocks(this);
-  }
+  private String serviceName = "apiary";
+  private String appId = "appid";
+  private String configPropertyPrefix = "prefix";
 
   @Test
   public void test_init() {
@@ -70,7 +60,7 @@ public class ApiaryRangerAuthAllAccessPolicyProviderTest {
         new ApiaryRangerAuthAllAccessPolicyProvider(rangerAdminRESTClient);
 
     apiaryRangerAuthAllAccessPolicyProvider.init(serviceName, appId, configPropertyPrefix);
-    verify(rangerAdminRESTClient, times(1)).init(anyString(), anyString(), anyString());
+    verify(rangerAdminRESTClient, times(1)).init(serviceName, appId, configPropertyPrefix);
 
     ServicePolicies servicePolicies = apiaryRangerAuthAllAccessPolicyProvider.getServicePolicies();
     assertThat(servicePolicies.getPolicies(), hasSize(1));
@@ -91,18 +81,16 @@ public class ApiaryRangerAuthAllAccessPolicyProviderTest {
     testImpl.init(serviceName, appId, "ranger.plugin.hive");
     ServicePolicies jsonPolicies = testImpl.getServicePoliciesIfUpdated(1L, 1L);
 
-    doReturn(jsonPolicies).when(rangerAdminRESTClient).getServicePoliciesIfUpdated(anyLong(), anyLong());
-
     ApiaryRangerAuthAllAccessPolicyProvider apiaryRangerAuthAllAccessPolicyProvider =
-        new ApiaryRangerAuthAllAccessPolicyProvider(rangerAdminRESTClient);
+        new ApiaryRangerAuthAllAccessPolicyProvider(testImpl);
 
     apiaryRangerAuthAllAccessPolicyProvider.init(serviceName, appId, configPropertyPrefix);
 
     ServicePolicies servicePolicies = apiaryRangerAuthAllAccessPolicyProvider.
         getServicePoliciesIfUpdated(1L, 1L);
 
-    // validate that we are *not* getting back the policies from the mocked JSON policy file, and that we *are*
-    // getting back the service definition from the mocked JSON policy file.
+    // validate that we are *not* getting back the policies from the test JSON policy file, and that we *are*
+    // getting back the service definition from the test JSON policy file.
 
     assertThat(jsonPolicies.getPolicies(), hasSize(10));
     assertThat(servicePolicies.getPolicies(), hasSize(1));
@@ -116,10 +104,8 @@ public class ApiaryRangerAuthAllAccessPolicyProviderTest {
     testImpl.init(serviceName, appId, "ranger.plugin.hive");
     ServicePolicies jsonPolicies = testImpl.getServicePoliciesIfUpdated(1L, 1L);
 
-    doReturn(jsonPolicies).when(rangerAdminRESTClient).getServicePoliciesIfUpdated(anyLong(), anyLong());
-
     ApiaryRangerAuthAllAccessPolicyProvider apiaryRangerAuthAllAccessPolicyProvider =
-        new ApiaryRangerAuthAllAccessPolicyProvider(rangerAdminRESTClient);
+        new ApiaryRangerAuthAllAccessPolicyProvider(testImpl);
 
     apiaryRangerAuthAllAccessPolicyProvider.init(serviceName, appId, configPropertyPrefix);