+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 4.2.0 - TBD
+### Added
+- `apiary-ranger-metastore-plugin` module has an `ApiaryRangerAuthAllAccessPolicyProvider` class that can be configured in ranger-hive-security.xml to allow for an audit-only Ranger pre-event listener. Intent is to configure it on the Apiary read-only HMS instances.
+
+
 ## 4.1.0 - 2019-06-27
 ### Added
 - `privileges-grantor-lambda` supports granting access to tables that were renamed.